pyproject.toml
[tool.poetry]
name = "gateway-service"
version = "0.1.0"
description = ""
authors = ["Danil Emurashin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
uvicorn = "^0.26.0"
pydantic-settings = "^2.1.0"
aiohttp = "^3.9.1"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
fastapi-pagination = "^0.12.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
